Adding Check On Users According to Role Permission

+++ packages/Webkul/Admin/src/DataGrids/Settings/UserDataGrid.php
@@ -124,22 +124,26 @@ class UserDataGrid extends DataGrid
      */
     public function prepareActions()
     {
-        $this->addAction([
-            'icon'   => 'icon-edit',
-            'title'  => trans('admin::app.settings.users.index.datagrid.edit'),
-            'method' => 'GET',
-            'url'    => function ($row) {
-                return route('admin.settings.users.edit', $row->user_id);
-            },
-        ]);
+        if (bouncer()->hasPermission('settings.users.users.edit')) {
+            $this->addAction([
+                'icon'   => 'icon-edit',
+                'title'  => trans('admin::app.settings.users.index.datagrid.edit'),
+                'method' => 'GET',
+                'url'    => function ($row) {
+                    return route('admin.settings.users.edit', $row->user_id);
+                },
+            ]);
+        }
 
-        $this->addAction([
-            'icon'   => 'icon-delete',
-            'title'  => trans('admin::app.settings.users.index.datagrid.delete'),
-            'method' => 'DELETE',
-            'url'    => function ($row) {
-                return route('admin.settings.users.delete', $row->user_id);
-            },
-        ]);
+        if (bouncer()->hasPermission('settings.users.users.delete')) {
+            $this->addAction([
+                'icon'   => 'icon-delete',
+                'title'  => trans('admin::app.settings.users.index.datagrid.delete'),
+                'method' => 'DELETE',
+                'url'    => function ($row) {
+                    return route('admin.settings.users.delete', $row->user_id);
+                },
+            ]);
+        }
     }
 }
