+++ packages/Webkul/Admin/src/Resources/views/settings/users/index.blade.php
@@ -51,9 +51,12 @@
                 src="{{ route('admin.settings.users.index') }}"
                 ref="datagrid"
             >
+                @php
+                    $hasPermission = bouncer()->hasPermission('settings.users.users.edit') || bouncer()->hasPermission('settings.users.users.delete');
+                @endphp
                 <!-- DataGrid Header -->
-                <template #header="{ columns, records, sortPage, applied}">
-                    <div class="row grid grid-cols-6 grid-rows-1 gap-[10px] items-center px-[16px] py-[10px] border-b-[1px] border-gray-300 text-gray-600 bg-gray-50 font-semibold">
+                <template #header="{columns, records, sortPage, applied}">
+                    <div class="row grid grid-cols-{{ $hasPermission ? '6' : '5' }} grid-rows-1 gap-[10px] items-center px-[16px] py-[10px] border-b-[1px] border-gray-300 text-gray-600 bg-gray-50 font-semibold">
                         <div
                             class="flex gap-[10px] cursor-pointer"
                             v-for="(columnGroup, index) in ['user_id', 'user_name', 'status', 'email', 'role_name']"
@@ -84,9 +87,11 @@
                         </div>
 
                         <!-- Actions -->
-                        <p class="flex gap-[10px] justify-end">
-                            @lang('admin::app.components.datagrid.table.actions')
-                        </p>
+                        @if ($hasPermission)
+                            <p class="flex gap-[10px] justify-end">
+                                @lang('admin::app.components.datagrid.table.actions')
+                            </p>
+                        @endif
                     </div>
                 </template>
 
@@ -95,7 +100,7 @@
                     <div
                         v-for="record in records"
                         class="row grid gap-[10px] items-center px-[16px] py-[16px] border-b-[1px] border-gray-300 text-gray-600 transition-all hover:bg-gray-50"
-                        style="grid-template-columns: repeat(6, 1fr);"
+                        :style="'grid-template-columns: repeat(' + (record.actions.length ? 6 : 5) + ', 1fr);'"
                     >
                         <!-- Id -->
                         <p v-text="record.user_id"></p>
@@ -146,23 +151,27 @@
 
                         <!-- Actions -->
                         <div class="flex justify-end">
-                            <a @click="id=1; editModal(record.user_id)">
-                                <span
-                                    :class="record.actions['0'].icon"
-                                    class="cursor-pointer rounded-[6px] p-[6px] text-[24px] transition-all hover:bg-gray-200 max-sm:place-self-center"
-                                    :title="record.actions['0'].title"
-                                >
-                                </span>
-                            </a>
+                            @if (bouncer()->hasPermission('settings.users.users.edit'))
+                                <a @click="id=1; editModal(record.user_id)">
+                                    <span
+                                        :class="record.actions['0'].icon"
+                                        class="cursor-pointer rounded-[6px] p-[6px] text-[24px] transition-all hover:bg-gray-200 max-sm:place-self-center"
+                                        :title="record.actions['0'].title"
+                                    >
+                                    </span>
+                                </a>
+                            @endif
 
-                            <a @click="deleteModal(record.actions['1']?.url, record.user_id, records?.length)">
-                                <span
-                                    :class="record.actions['1'].icon"
-                                    class="cursor-pointer rounded-[6px] p-[6px] text-[24px] transition-all hover:bg-gray-200 max-sm:place-self-center"
-                                    :title="record.actions['1'].title"
-                                >
-                                </span>
-                            </a>
+                            @if (bouncer()->hasPermission('settings.users.users.delete'))
+                                <a @click="deleteModal(record.actions['1']?.url, record.user_id, records?.length)">
+                                    <span
+                                        :class="record.actions['1'].icon"
+                                        class="cursor-pointer rounded-[6px] p-[6px] text-[24px] transition-all hover:bg-gray-200 max-sm:place-self-center"
+                                        :title="record.actions['1'].title"
+                                    >
+                                    </span>
+                                </a>
+                            @endif
                         </div>
                     </div>
                 </template>
